Remove second mocks file for Accessibility
